Begun challenge
I have begun work on the challenge, I have started by creating a new class which will take a string and separate it into its component parts as long as the separator is defined for example ' '.

Next, I need to loop through the separated vector to process the cannibal numbers.